ASoC: SOF: fix Kconfig dependencies
The existing module dependencies are handled in the wrong way.  We
need to select first acpi/pci, then platforms and last the core.
Doing otherwise would result in broken configurations where acpi/pci
stuff is compiled as built-in and platforms or SOF core as a module.

Rebuild Kconfig to use 'select' statements in the right order. The
platform selections are now exposed with booleans only. The choice of
module or built-in can now happen only at the pci/acpi/spi level, the
dependencies will be compiled accordingly. As a result, "make
menuconfig" does not expose how dependencies are handled.

Corner cases:
PCI is required for sof-pci-dev, devm_pci_enable is not defined otherwise.
ACPI and SOF_INTEL_TOPLEVEL can be enabled for COMPILE_TEST

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>